Add an apply function
Meteor connections always have an apply function next to the call function. The call function in SimpleDDP has almost the same syntax as apply in Meteor. So to get a little closer to Meteor compatibilty, we could simply use the call function as an apply function. Only difference is that you can't pass it the apply options as mentioned in https://docs.meteor.com/api/methods.html#Meteor-apply